Review fixes (#9)
* Review notes pt1

* pt2

* pt3

* add 2d nonce callout

* pt4

* initial review fixes

* second round review fixes

* remove account from AOAConfig

* more fixes

* use full config bytes in binding

* more fixes, combine validity window and execute

* remaining review comments

---------

+++ src/policies/accounting/RecurringAllowance.sol
@@ -16,19 +16,19 @@ library RecurringAllowance {
         /// @dev Maximum spend per period window.
         uint160 allowance;
         /// @dev Period length in seconds.
-        uint48 period;
+        uint40 period;
         /// @dev Start timestamp (seconds) inclusive.
-        uint48 start;
+        uint40 start;
         /// @dev End timestamp (seconds) exclusive.
-        uint48 end;
+        uint40 end;
     }
 
     /// @notice Stored usage snapshot for a particular active period.
     struct PeriodUsage {
         /// @dev Period start timestamp (seconds).
-        uint48 start;
+        uint40 start;
         /// @dev Period end timestamp (seconds).
-        uint48 end;
+        uint40 end;
         /// @dev Amount spent during the period window.
         uint160 spend;
     }
@@ -43,13 +43,28 @@ library RecurringAllowance {
     ///                         Errors                           ///
     ////////////////////////////////////////////////////////////////
 
+    /// @notice Thrown when the period length is zero.
     error ZeroPeriod();
+
+    /// @notice Thrown when the allowance is zero.
     error ZeroAllowance();
-    error InvalidStartEnd(uint48 start, uint48 end);
-    error BeforeStart(uint48 currentTimestamp, uint48 start);
-    error AfterEnd(uint48 currentTimestamp, uint48 end);
+
+    /// @notice Thrown when `start >= end` in the limit bounds.
+    error InvalidStartEnd(uint40 start, uint40 end);
+
+    /// @notice Thrown when the current timestamp is before the allowance window.
+    error BeforeStart(uint40 currentTimestamp, uint40 start);
+
+    /// @notice Thrown when the current timestamp is at or past the allowance window end.
+    error AfterEnd(uint40 currentTimestamp, uint40 end);
+
+    /// @notice Thrown when the spend value is zero.
     error ZeroValue();
+
+    /// @notice Thrown when cumulative spend exceeds `type(uint160).max`.
     error SpendValueOverflow(uint256 value);
+
+    /// @notice Thrown when cumulative spend exceeds the period allowance.
     error ExceededAllowance(uint256 value, uint256 allowance);
 
     ////////////////////////////////////////////////////////////////
@@ -109,7 +124,7 @@ library RecurringAllowance {
         view
         returns (PeriodUsage memory)
     {
-        uint48 currentTimestamp = uint48(block.timestamp);
+        uint40 currentTimestamp = uint40(block.timestamp);
         if (currentTimestamp < limit.start) revert BeforeStart(currentTimestamp, limit.start);
         if (currentTimestamp >= limit.end) revert AfterEnd(currentTimestamp, limit.end);
 
@@ -118,10 +133,10 @@ library RecurringAllowance {
         bool lastStillActive = currentTimestamp < lastUpdated.end;
         if (lastExists && lastStillActive) return lastUpdated;
 
-        uint48 currentPeriodProgress = (currentTimestamp - limit.start) % limit.period;
-        uint48 start = currentTimestamp - currentPeriodProgress;
+        uint40 currentPeriodProgress = (currentTimestamp - limit.start) % limit.period;
+        uint40 start = currentTimestamp - currentPeriodProgress;
         bool endOverflow = uint256(start) + uint256(limit.period) > limit.end;
-        uint48 end = endOverflow ? limit.end : start + limit.period;
+        uint40 end = endOverflow ? limit.end : start + limit.period;
         return PeriodUsage({start: start, end: end, spend: 0});
     }
 }
